Update patch set 1
Patch Set 1:

Sorry, I didn't mean to disrespect your right for proper attribution. I just wanted the asm parts split off from the rest of the (already large, and already in review for a long time) CL 19336 and this seemed the fastest way to do so.

In any case, I read the asm discussion on golang-dev as concluding that there's no hurry on landing these asm functions. I'm happy to abandon this CL, and for you to make a separate one in the future, if we do decide to add asm.

Patch-set: 1
Reviewer: Gerrit User 5899 <5899@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0