Update patch set 1
Patch Set 1:

what's the maximum length of the string that is
duplicated?

I'd imagine that the max length to be fairly small as the
longer the string, the less likely for it to be used in
multiple packages. (Unless it's a const string, but in
that case, I'd argue it's better to name the string according
to the name of the fully qualified name of the const).

If my assumption is right, then we could just raise
the current limit on content addressable string length,
name the string after its const name, and ignore long
anonymous strings.

Patch-set: 1
Reviewer: Gerrit User 5055 <5055@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0