Update patch set 18
Patch Set 18:

Nigel and Joe seem happy. Russ should be happy now that you've addressed his remaining requests.

I'm going to submit this while the trybots are happy so this doesn't rot, especially since Klaus is going to be away for a bit. This review has gone on long enough, as Russ said.

Any problems with this CL can be address in follow-up CLs.

Patch-set: 18
Label: Code-Review=0