Update patch set 1
Patch Set 1:

I appreciate the amount of work here.

There is an argument that says that every little helps, but the history of partial fixes to timing leaks is that they're repeatedly found to be ineffective. At the moment, at least, we have a clear story: CBC modes are broken and are supported only for legacy compatibility.

This change, no doubt, reduces the timing leak, but things like MAC extraction haven't been made constant-time (and it's hard and slow to do that). There are some other probable small leaks here too. It realistically takes a few days with a test suite and lots of RDTSC instructions to have confidence in a constant-time implementation of something complex like CBC decoding.

I'm open to being persuaded that "every little helps", but I'm skeptical that we would be able to prevent the attack and I don't think that we should remove the warning.

Patch-set: 1