Update patch set 5
Patch Set 5:

(6 comments)

Maybe it is just me, but I find this code terribly confusing.  I think it would help a lot to add some comments at the top describing invariants in the representation, and maybe an example.
It might also help clarify things to use separate maps for the boolean and the range facts.
The facts table contains a slice of facts known about each pair.  How many entries are in this slice?  When do they get added/deleted?  My impression is that there is one entry per dominator which branches on a fact relating to the two values keying that fact table entry.
But I'm not sure, some explanation would be helpful.

Patch-set: 5
Label: Code-Review=0