runtime: preserve darwin/arm{,64} callee-save registers
CL #14603 attempted to preserve the callee-save registers for
the darwin/arm runtime initialization routine, but I believe it
wasn't sufficient and resulted in the crash reported in issue

Saving and restoring the registers on the stack the same way
linux/arm does seems more obvious and fixes #14778, so do that.

Even though #14778 is not reproducible on darwin/arm64, I applied
a similar change there.
Someone more knowledgeable in asm need to verify the correctness;
linux/arm64 does not preserve the registers on the stack so there
is no precedent.

Note that if the darwin/arm64 changes are correct, linux/arm64 and
other arm64 platforms should probably be updated in a later CL.

Finally, this CL is a candidate for a 1.6 minor release for the same
reason CL #14603 was in a 1.5 minor release (as CL #16968). It is
small and only touches the iOS platforms and gomobile on darwin/arm
is currently useless without it.

Fixes #14778
Fixes #12590 (again)

Change-Id: I7401daf0bbd7c579a7e84761384a7b763651752a