Update patch set 1
Patch Set 1:

(1 comment)

I just tried and failed to reproduce the crash in #14778 with Go 1.5.3, confirming David's observation that the register-clobbing runtime·libpreinit call was added late in the Go 1.6 cycle.

That means that there is a released version of Go that works with current gomobile darwin/arm. It also means that #14778 is a regression from 1.5.

Patch-set: 1