Update patch set 1
Patch Set 1:

This may changed the struct returned by reflect.Type.Method when called on an unexported method. I'm still trying to work out if those changes ever cause a practical problem (there's a shortage of code depending on it, and the docs don't really define the value of those fields when the method is unexported).

Patch-set: 1