Update patch set 5
Patch Set 5:

(9 comments)

(This is only a partial review.)

In the gold linker I found that on a 32-bit host the linker really did run of memory when it mmapped all the input files, even though the resulting program could still run on a 32-bit system.  We should be really sure that we don't care about building large programs on 32-bit hosts, or we should adjust the code to still work by reading from a file and use that code on 32-bit hosts.

Patch-set: 5
Label: Code-Review=0