Update patch set 9
Patch Set 9:

> My motivation is entirely empirical.

Definitely. To be clear, I think the current 6% improvement is awesome, and worth pursuing regardless.

I meant that benchmarking mmap vs slurping would be interesting to measure as a followup effort. Since switching between the two should be localized to objReader, the rest of this CL is still very useful IMO.

Patch-set: 9
Label: Code-Review=0