Update patch set 10
Patch Set 10: -Run-TryBot

Unfortunately following the review suggestions has slowed down the code so now this CL is a wash.

I'll try backing out the MAP_SHARED change tomorrow, but linker performance is a diversion from my binary size goals, so if it isn't easy to get back to the -6% I was seeing at the beginning I'll leave this for someone else.

Patch-set: 10
Label: Run-TryBot=0