Update patch set 3
Patch Set 3:

It's not obvious to me that slash-splitting the pattern is the right approach. Today it's a single regular expression, and it would be simpler if it could continue to be. Here are the constraints I see.

1. We want -run=Foo/bar to run Foo in order to get to Foo/bar if the latter exists.

2. Since we can't know whether Foo/bar actually exists without running Foo, -run=Foo/bar must run Foo even if Foo/bar doesn't exist. Strange but apparently unavoidable.

3. We might want -run=bar to run Foo/bar, but since that would imply running Foo, it would imply running everything, which is too confusing. So something about the pattern has to match the top-level test name.

The current docs for the -run match (from go help testflag) are:

	-run regexp
	    Run only those tests and examples matching the regular
	    expression.

The smallest semantic change that works, based on the above constraints, seems to be:

	-run regexp
	    Run only those tests, subtests, and examples with names
	    or potential subtest names matching the regular expression.
	    To match a subtest name, the matched text must begin in the
	    top-level test name (before the first slash).

This seems significantly easier to explain, and a significantly smaller departure from past behavior, than interpreting the pattern as a /-separated list of regexps.

Of course the regexp package today doesn't help implement that change, but we can fix that.

What do you think of the semantics? Are there other constraints you see?

Thanks.

Patch-set: 3
Label: Code-Review=0