cmd/compile, etc: drop *typeAlg from ptrType
There are many pointer types. They all have the same predictable
algorithm table. So do a quick check before using the *typeAlg
at runtime.

One potentially hot path for accessing the algorithm table is map
lookups and assignments. To avoid the extra check on that path, this
CL adds the key *typeAlg to the mapType structure.

Shrinks binaries:

	cmd/go: -143KB (1.4%)
	jujud:  -2.3MB (3.6%)

For #6853.

Change-Id: I814e1e94418437b7595e78034a0171f56af17f62