Update patch set 1
Patch Set 1:

I'm not saying the sleep needs to stay. I'm saying you need to address the reason it is there, namely to avoid the excessive CPU and overall execution time caused by lightweight use of goroutines otherwise. Dismissing those programs or presenting other synthetic benchmarks does not address them. If you can't address them directly, the best argument to take the hit would be that in _real_ programs it causes  a significant improvement. And again I think it's a mistake to spend 100% extra CPU making the program 25% faster. That is an overall 50% waste of CPU (100 units for 100% x 1.0, 150 units for 200% x 0.75).

Patch-set: 1