Update patch set 1
Patch Set 1:

> I'm not saying the sleep needs to stay. I'm saying you need to
 > address the reason it is there, namely to avoid the excessive CPU
 > and overall execution time caused by lightweight use of goroutines
 > otherwise. Dismissing those programs or presenting other synthetic
 > benchmarks does not address them. If you can't address them
 > directly, the best argument to take the hit would be that in _real_
 > programs it causes  a significant improvement. And again I think
 > it's a mistake to spend 100% extra CPU making the program 25%
 > faster. That is an overall 50% waste of CPU (100 units for 100% x
 > 1.0, 150 units for 200% x 0.75).


It's halting problem. We don't know what a goroutine will be doing in future. And so we can't decide whether it makes sense to stash a goroutine locally or not. We can only choose between lower latency/higher CPU load and higher latency/lower CPU load. In this context I would choose eager goroutine execution, because user has control over granularity and communication patterns but don't have control over the sleep. Just any program can't be magically executed in the most efficient way, user needs to cooperate (e.g. compilers generally don't replace linked lists with arrays).

We could involve compiler so that it detects chan operations that go in short succession and pass hints to runtime. Hint to the first chan operation will allow it to stash a goroutine locally. Hint to the second operation instructs to make the goroutine available to other threads if the operation is not going to block. But this is fragile and I don' think that the complexity worth it for the sake of synthetic benchmarks.

Re CPU load, it's not that bad if you look at tests with more local work. And in the end it's users choice (again, he has control over it).

Patch-set: 1