Update patch set 6
Patch Set 6:

(1 comment)

I think the proposed semantics results in very inconvenient usage.

Use case: I want to run all direct subtests of TestRoundEven and TestRoundOdd with a name containing "foo".
With my proposed semantics, this would be
    go test --run=Round/foo

For your proposal, I would have to write:
    go test --run=Round[^/]*/[^/]*foo

It would be very tempting to write --run=Round.*/.*foo, or even --run=Round.*foo, but all uses of ".*" would always match all tests if the preceding pattern has been matched. I think this is both very confusing and very inconvenient.

Perhaps the following would be an acceptable explanation for the run flag:

 -run regexp_list
 Run only those tests and examples matching the regular
 expression. Additional regular expressions my be given,
 separated by a '/', to filter corresponding subtests, if any.

There are some subtleties in my CL to handle cases where the subtest name contains a '/', but even that would work. We cold further restrict the use of '/' in a regexp pattern.

Alternative approach:
- run pattern
 Run only those tests and examples matching the regular
expression given by pattern. If pattern contains a '/' it is
treated as a glob that can also filter subtests.

In this case, I would have to write the above search as:

    go test --run=TestRound*/*foo*

This has its advantages, but also its weirdnesses. As it is less convenient than the regexp per level method, I favor the original proposal.

There is also the slight issue of having to modify regexp and to cope with the fact that RunTests and RunExamples are not documented as internal, but take a matchstring function that is supposed to be a regular expression matcher.

 > It's not obvious to me that slash-splitting the pattern is the
 > right approach. Today it's a single regular expression, and it
 > would be simpler if it could continue to be. Here are the
 > constraints I see.
 > 
 > 1. We want -run=Foo/bar to run Foo in order to get to Foo/bar if
 > the latter exists.
 > 
 > 2. Since we can't know whether Foo/bar actually exists without
 > running Foo, -run=Foo/bar must run Foo even if Foo/bar doesn't
 > exist. Strange but apparently unavoidable.
 > 
 > 3. We might want -run=bar to run Foo/bar, but since that would
 > imply running Foo, it would imply running everything, which is too
 > confusing. So something about the pattern has to match the
 > top-level test name.
 > 
 > The current docs for the -run match (from go help testflag) are:
 > 
 > -run regexp
 > Run only those tests and examples matching the regular
 > expression.
 > 
 > The smallest semantic change that works, based on the above
 > constraints, seems to be:
 > 
 > -run regexp
 > Run only those tests, subtests, and examples with names
 > or potential subtest names matching the regular expression.
 > To match a subtest name, the matched text must begin in the
 > top-level test name (before the first slash).
 > 
 > This seems significantly easier to explain, and a significantly
 > smaller departure from past behavior, than interpreting the pattern
 > as a /-separated list of regexps.
 > 
 > Of course the regexp package today doesn't help implement that
 > change, but we can fix that.
 > 
 > What do you think of the semantics? Are there other constraints you
 > see?
 > 
 > Thanks.

Patch-set: 6