Update patch set 4
Patch Set 4:

(3 comments)

What are your thoughts on just doing a busy-loop here for 3 microseconds, possibly with an osyield in there? Something like:

begin := nanotime()
osyield()
for nanotime() - begin < 3e3 { }

My thinking is that a 3us sleep is not useful because it gives up the CPU for 3us. It's useful because it does one of two things: it slightly delays the steal, or if there's actually work on other threads, they get to run (and we sleep much longer). The above snippet should accomplish these two goals regardless of how bad the OS's sleep is, assuming its nanotime isn't equally broken. And if the yielding part doesn't matter and it's only the slight delay that matters, we could leave out the yield.

Patch-set: 4
Label: Code-Review=0