Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> Well that didn't work. Curious that it's failing on all of the 32 bit arches.

Russ and I found what's probably the problem. typedmemmove and friends call heapBitsBulkBarrier, which *does* care about the destination and is a no-op if dst is a global. Presumably I'm seeing the difference in SSA just because it isn't calling typedmemmove in as many cases as the old backend.

A possible solution is to make heapBitsBulkBarrier understand the data/bss bitmap and perform the appropriate individual barriers.

Patch-set: 1