fmt: cleanup general reflect value handling
This CL focuses on simplifying code around the big reflect switch
in printReflectValue that deals with using reflect to print values.
Followup CLs will focus on further cleanup and improvement of the
reflect switch itself.

Merge printReflectValue into printValue. Determine if handleMethods
was already called in printArg by checking if depth is 0. Do not
call handleMethods on depth 0 again in printValue to not introduce
a performance regression. handleMethods is called already in printArg
to not introduce a performance penalty for top-level Stringer,
GoStringer, Errors and Formatters by using reflect.ValueOf on them
just to retrieve them again as interface{} values in printValue.

Fix a bug that would incorrectly print the whole map instead of
just the value for a key if the returned value by the map for
the key was an invalid reflect value.

Remove p.value and instead use p.arg to store reflect values to be
printed in case an unsupported verb is encountered. Always unpack
reflect values to print the unsupported verb error string similar
to how normal printing handles reflect values. This changes the
behavior to print the real type of the value wrapped in a
reflect.Value if the value was supplied to an unsupported verb
instead of printing "reflect.Value=" inside the error string.

The removal of p.value has a negative effect on performance due
to struct alignment. We can revisit this in an extra CL after any
other followup CL to find a new sweetspot that regains the performance
introduced by the last change to the struct in CL 20838.

Remove getField helper function that would provide a fast path to
handle interface fields instead of calling printValue directly on
the field. Use a direct call to printValue for consistency with other
types such as maps, arrays, slices and pointers instead. A later CL
will focus on speeding up interface handling so that all these types
can benefit from such an optimization.

Cleanup and consolidate the code of reflect.Ptr case to avoid
the fallthrough and BigSwitch label.

name               old time/op  new time/op  delta
SprintfStringer-2   367ns ± 2%   383ns ± 3%  +4.20%  (p=0.000 n=20+19)

Change-Id: Ib9a39082ed1be0f1f7499ee6fb6c9530f043e43a