Update patch set 4
Patch Set 4:

(2 comments)

> (3 comments)
 > 
 > What are your thoughts on just doing a busy-loop here for 3
 > microseconds, possibly with an osyield in there? Something like:
 > 
 > begin := nanotime()
 > osyield()
 > for nanotime() - begin < 3e3 { }
 > 
 > My thinking is that a 3us sleep is not useful because it gives up
 > the CPU for 3us. It's useful because it does one of two things: it
 > slightly delays the steal, or if there's actually work on other
 > threads, they get to run (and we sleep much longer). The above
 > snippet should accomplish these two goals regardless of how bad the
 > OS's sleep is, assuming its nanotime isn't equally broken. And if
 > the yielding part doesn't matter and it's only the slight delay
 > that matters, we could leave out the yield.


I agree with your reasoning. I don't have a good answer. But here are some points:

The usleep still keeps CPU load as reported by getrlimit on 125% instead of 200%. I assume if we use active spinning, CPU load will be 200%. I would not be surprised that the CPU is actually running all the time effectively on behalf of our process, but only 25% of that time is accounted to our process. But these accounting numbers are what people pay great attention to (either when they monitor prod systems, or then they compare different languages on synthetic benchmarks and then loudly scream about inefficiency of one or another).

I think that yielding to other processes is not a bad thing. These other processes still need to do their work sooner or later. And this looks like a good point to yield. Note that it should matter only on an overloaded system, because otherwise either our or other processes will be moved to an idle CPU.

Regardless of sleep/spin what I think would be an improvement is: remember what Ps have runnext, values of runnext and when we observed it. Then continue polling/sleeping/spinning, but start stealing runnext only after it has been there for more than 3us. Currently we wait for 3us for each individual P. This can lead to a very substantial latency increase if we hit an unfortunate sequence of observing runnext, sleeping, discovering that it has gone and then repeating it for 62 other Ps. With the current 100 us sleep, we can sleep for up to 6.3ms with GOMAXPROCS=64. What I am proposing is to sleep to 3 us once and then start stealing runnext's.

Active spinning is more useful if we do the previous point, because we can poll other Ps while we are waiting for runnext's.

But I would prefer to not do anything of that in this change. This change is conservative and looks like a strict improvement that fixes two real problems (99% latency reduction on windows and 15% HTTP latency-99 reduction).

Patch-set: 4
Label: Code-Review=0