Update patch set 1
Patch Set 1:

(2 comments)

Amazing that building an intermediate slice is faster.

cmd/compile went through a project recently of replacing its Node linked lists with slices. It would be nice to do something similar in cmd/link, and then delete functions like listsort.

If you're up for continuing this work, the next step after this would be to convert the entire dodata routine to working off slices, and then build it just before it is called.

Patch-set: 1
Label: Code-Review=0