Update patch set 2
Patch Set 2:

> I would prefer to find that "sweet spot" so I am not reviewing a CL
> that slows down a very common operation.

Would it be ok for now to leave an unused field as filler in the struct with a comment? 

I have not yet been able yet to find a reordering on my intel haswell that does not regress performance back to a level before https://go-review.googlesource.com/#/c/20838/.

I will have a look with another CPU and change the CL to not cause a slow down either way.

Patch-set: 2