Update patch set 1
Patch Set 1:

Thanks for doing this. Would you be able to split this into two CLs? On for test changes and one for non test changes. I think the latter should be uncontroversial and should merge quickly. The amount of dead test code concerns me and I would like to investigate it separately because the dead code might be a sign that something has fallen out of coverage and rather than deleting it code maybe it should be revived. I don't know the answer to that yet. But hopefully with a smaller change it should be possible to answer. 

Thanks

Dave

Patch-set: 1