Update patch set 2
Patch Set 2:

I find the two forms pretty close in readability. And other areas of this code--and the rest of the compiler--have very, very serious readability problems, which I try to incrementally fix as I work on it. I will be delighted if/when binary export makes all this code obsolete. In the meantime, I saw a small, straightforward, obviously correct modification with minor performance benefit. If nothing else, it will make other performance opportunities in the compiler clearer in the short term. And for the record, I also tried out a more serious modification that set up byte buffers much higher up in the fmt.go stack; it was far less readable and yielded worse results.

Patch-set: 2