Update patch set 2
Patch Set 2:

> Padding would work but i suspect a reordering would do fine.

I found that it is the (cache) size. With reflect the pp struct is exactly 3 cache lines with 64 byte. Removing reflect.Value shrunk it to 168 bytes. To counter this i increased internal buffer in fmt to 96 bytes so it ends up at 192 byte again. Which should be better than wasting that space on padding. This recovers the performance loss. Will update the test cases/comments for overflow of intbuf and adjust the cl.

Patch-set: 2