Update patch set 2
Patch Set 2:

For the record, my intention is figuring out why such string
manipulations are in the hot code path in the compiler in
the first place.

And I offered a guess, and I'm interested to see if it's true
and whether we can do something about it. (Unfortunately
I currently lacked the required time to pursue this issue.)

Patch-set: 2