Update patch set 2
Patch Set 2:

> >
 > > Please upload issue14790.test.exe and a somewhere.
 > 
 > I have used my single cpu windows 7 again. I have create a github
 > repo with those files. See https://github.com/alexbrainman/goissue14790/blob/master/20160322/transcript.txt
 > for details.
 


The GOMAXPROCS=1 case is working as intended, it has latency of an order of 1ms which is the time to retake the single P from the blocked syscalls.

Please upload a trace for GOMAXPROCS>1 case. I think we don't need more than 1 CPU on the machine, we just need GOMAXPROCS=4. Because once WFSO blocks on a thread, OS will schedule another thread.

Patch-set: 2