Update patch set 1
Patch Set 1:

The CL description says what is happening, but not why. How will the compiler use position information in export data?

And if this is not for the compiler, what is the effect on compiler performance?

(I am hoping the binary export data format helps recover some of the compiler speed from 1.4.)

Patch-set: 1
Label: Code-Review=0