Update patch set 4
Patch Set 4:

(5 comments)

> (5 comments)
 > 
 > I don't actually know whether it's worth doing a piecemeal redesign
 > of the object file format.

I think many of the necessary changes to the object format wouldn't make much sense in a design document. They are small changes meant to optimize the parsing stage which takes up about 40% of link time. The best way forward would seem to me coding the changes and proving they are worthwhile, backing it up by numbers.
Also the fact that the format is non-standard seems to me like it should give the flexibility to make changes to it.
  
If you disagree I could try to write something up.

Patch-set: 4
Label: Code-Review=0