Update patch set 1
Patch Set 1:

> Does this affect the shared library execution mode(s) at all?

I do not believe so. I haven't put any effort into optimizing buildmode=shared yet, so in deadcodepass.init all symbols are marked as reachable if buildmode=shared. For the other build modes we have a main and so this analysis works as it should in a normal binary.

(One of the nice things deadcode does is prune used methods that are inlined. When I'm finished mucking about in the type info layout I'll try to make that work for buildmoe=shared.)

Patch-set: 1