Update patch set 1
Patch Set 1: Code-Review+1

LGTM. This is great. Now that the linker is removing these, the compiler should too (slightly different conditions). The linker is saying if method x exists but no interface being kept in the binary mentions x, then drop x. The compiler can now say if method x exists but no interface in the current package (anywhere) mentions x, then drop x. If a package defines an interface that turns out not to be needed in a particular binary, then the linker drop is the backstop, but the compiler drop will shorten the tables and in general compiler work is preferred to linker work because it parallelizes and is reused more easily.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1