Update patch set 3
Patch Set 3:

I think a useful test would be a C program linked with a Go archive, where the C program dereferences a NULL pointer or otherwise crashes with a signal.  We want to verify that that the program does in fact crash at that point--it does not continue executing--and that the exit status indicates that it crashed with the appropriate signal.

It's true that such a test ought to pass before your change as well, but your change is going to rely on a more subtle behavior--that continuing execution will trigger the same signal again--so I think a test would be a good idea.

Patch-set: 3
Label: Code-Review=0