cmd/compile: extend prove pass to handle constant comparisons
Find comparisons to constants and propagate that information
down the dominator tree.  Use it to resolve other constant
comparisons on the same variable.

So if we know x >= 7, then a x > 4 condition must return true.

This is only a partial implementation right now, just enough
to prove that it works.  It allows us to use "_ = b[7]" hints
to eliminate bounds checks.

DO NOT SUBMIT

Change-Id: Idbf230bd5b7da43de3ecb48706e21cf01bf812f7