Update patch set 1
Patch Set 1:

"Of course next request will be _ = b[:i+7] "

The way I dealt with things like "i+7" in the too-slow-signed-only inference was to record all (in)equalities as   "a <= b+K", with a special case for "a <= K".  The lower-bound inference for constants needed more work.

This worked, provided that you could correctly model overflow -- you need to know that b+K won't overflow.  Whenever you have a plan inequality, like a<b, you can replaces that with a<=b-1 and you know that b won't underflow (because a < b), but if you have t=a+K you don't know about lack of overflow unless you know something about the signs/sizes of t and a.

Patch-set: 1