Update patch set 6
Patch Set 6:

> (1 comment)
 > 
 > Do you have access to a GNU/Linux system for debugging?  I looked
 > at the test but I don't know what is going wrong.

Yeah, I've got some linux VMs that I can try this on.  I wonder if the signal handling in Linux is different somehow and the externally generated signals are not being retriggered.  I'll dig into it and think of a solution if that's the case.

Patch-set: 6
Label: Code-Review=0