Update patch set 8
Patch Set 8: -Run-TryBot

> Ah.. nope.  When I changed the external signal to SIGSEGV then the
 > tests fail.  The handler must be returning, but there's nothing to
 > retrigger the signal (since it was caused externally).  How do we
 > detect that?

Ah, right.  In that case we do not want to return to the handler, as the wrong thing will happen.  I think we need to change the call to badsignal to pass along the signal code; see sighandler in signal_amd64x.go for how to retrieve it.  Or maybe better than passing the sigcode would be to pass sigcode != _SI_USER.  Then pass that along to raisebadsignal, and then we know whether it is safe to return.

At least, that is my first thought, happy to hear other suggestions.

Patch-set: 8
Label: Run-TryBot=0