Update patch set 2
Patch Set 2:

The problem i had with this kind of benchmark is that it does not measure the impact of the branches on a real workload where its unlikely we just grow the same size of array a few million times in a row. This can easily offset the 7ns advantage for the GrowSliceStruct24Bytes-4. Each additional branch adds costs in general to any other size of element. That is why i avoided this in my original CL. However avoiding the maxcap division alltogether is nice.

Patch-set: 2
Label: Code-Review=0