Update patch set 4
Patch Set 4:

Here is what I suggest. The original code wasn't ideal either, as what it should have been doing was seeing if the final argument was a ...interface{}, and if so checking if the previous argument was of type string.

I suggest making that change, and if that fails to find a valid signature, run the new call-site code. There are cases where the call-site code will fail. Here's one: logger.Printf("WARNING", "%d", 3)

Patch-set: 4
Label: Code-Review=0